refactoring in progress
